+++ samples/openapi3/client/petstore/python/src/petstore_api/components/schema/array_of_number_only.py
@@ -15,7 +15,7 @@ Items: typing_extensions.TypeAlias = schemas.NumberSchema
 
 class ArrayNumberTuple(
     typing.Tuple[
-        typing.Union[float, int],
+        typing.Union[int, float],
         ...
     ]
 ):
