+++ samples/openapi3/client/petstore/python/src/petstore_api/components/schema/object_with_only_optional_props.py
@@ -29,7 +29,7 @@ class ObjectWithOnlyOptionalPropsDict(schemas.immutabledict[str, schemas.OUTPUT_
         ...
     
     @typing.overload
-    def __getitem__(self, name: typing_extensions.Literal["b"]) -> typing.Union[float, int]:
+    def __getitem__(self, name: typing_extensions.Literal["b"]) -> typing.Union[int, float]:
         ...
     
     def __getitem__(
