+++ samples/openapi3/client/petstore/python/src/petstore_api/paths/fake_refs_number/post/responses/response_200/__init__.py
@@ -12,7 +12,7 @@ from .content.application_json import schema as application_json_schema
 @dataclasses.dataclass
 class ApiResponse(api_response.ApiResponse):
     response: urllib3.HTTPResponse
-    body: typing.Union[float, int]
+    body: typing.Union[int, float]
     headers: schemas.Unset = schemas.unset
